DiJia/DiJia_Backend/fastapi_api/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openai==1.11.0
python-dotenv==1.0.1
pydantic